// src/types/components/calendar.ts
type Placement =
  | "right-start"
  | "left-start"
  | "right-end"
  | "left-end"
  | "top-start"
  | "top-end"
  | "bottom"
  | "bottom-start"
  | "bottom-end";

type PopoverPlacement = {
  placement: Placement;
  colNum: number;
  rowNum: number;
};